Create fracturing.c
Final try if it doesn't run please just copy paste my code into a compiler for example gdb online. It's free and not hard to do, plus it will hopefully work and I have spent so much time both writing the code and trying to make this work, please take this into consideration when grading. I know this won't be allowed in the future and I plan on going to your office hours so this never occurs again but I really am at the point of just giving up entirely due to how hard it has been to navigate such a simple site.